requirements.txt
fastapi>=0.115.0,<0.116
uvicorn>=0.30.6,<0.31
openai>=1.55.0,<2.0
python-dotenv>=1.0.1,<2.0
supabase>=2.7.4,<3.0
httpx>=0.27.2,<0.28

langchain>=0.3.27,<0.4
langchain-core>=0.3.76,<0.4
langchain-community>=0.3.30,<0.4
langchain-openai>=0.3.33,<0.4
langchain-text-splitters>=0.3.11,<0.4

langgraph>=0.6.7,<0.7
langgraph-prebuilt>=0.6.4,<0.7
langgraph-checkpoint>=2.1.1,<3.0
langgraph-sdk>=0.2.9,<0.3

langsmith>=0.4.31,<0.5
pytest>=8.4.2,<9.0
